requirements.txt
Django==4.2.19
psycopg2-binary==2.9.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
dj-rest-auth==7.0.1
django-allauth==65.4.1
django-tenants==3.7.0
django-storages==1.14.5
Pillow==10.4.0